requirements.txt
streamlit
spacy
scikit-learn
pandas
sentence-transformers
pdfminer.six
python-docx
